GOOWOO-681 addressing MORE comments
