+++ js/src/pages/markets/edit-market-modal/index.js
@@ -10,7 +10,7 @@ import { PRIMARY_MARKET_ID } from '../constants';
 import AppModal from '~/components/app-modal';
 import MarketForm from '../market-form';
 import MarketFields from '../market-fields';
-import EditMarketButtons from './edit-market-buttons';
+import ModalFooter from './modal-footer';
 
 /**
  * @typedef {import('~/data/actions').TargetAudienceData } TargetAudienceData
@@ -18,22 +18,7 @@ import EditMarketButtons from './edit-market-buttons';
  */
 
 /**
- * Event fired when the "Cancel" button in the EditMarketModal is clicked.
- * @event gla_cancel_button_clicked
- * @property {string} context The context in which the cancel button click happened, e.g. "edit_market_modal".
- */
-
-/**
- * Event fired when the "Save" button in the EditMarketModal is clicked.
- * @event gla_save_button_clicked
- * @property {string} context The context in which the save button click happened, e.g. "edit_market_modal".
- */
-
-/**
  * Modal component for editing an existing market.
- *
- * @fires gla_cancel_button_clicked when the cancel button is clicked with context of "edit_market_modal"
- * @fires gla_save_button_clicked when the save button is clicked with context of "edit_market_modal"
  *
  * @param {Object} props
  * @param {{ id: string, label: string }} props.market The market being edited.
@@ -66,7 +51,7 @@ const EditMarketModal = ( { market, targetAudience, onRequestClose } ) => {
 				onSubmit={ onRequestClose }
 			>
 				<MarketFields />
-				<EditMarketButtons onRequestClose={ onRequestClose } />
+				<ModalFooter onCancel={ onRequestClose } />
 			</MarketForm>
 		</AppModal>
 	);
